Add productDescription componemt
