+++ src/App.js
@@ -5,7 +5,47 @@ import Header from './app/components/Header';
 import Kids from './Pages/Kids';
 import Men from './Pages/Men';
 import Women from './Pages/Women';
+import clothImg from './assets/images/product.png'
 
+const cloth = [
+  {
+    id: 1,
+    brand: 'Apollo running shot',
+    amount: '$50.00',
+    product: clothImg
+  },
+  {
+      id: 2,
+      brand: 'Apollo running shot',
+      amount: '$50.00',
+      product: clothImg
+  },
+  {
+      id: 3,
+      brand: 'Apollo running shot',
+      amount: '$50.00',
+      product: clothImg
+  },
+  {
+      id: 4,
+      brand: 'Apollo running shot',
+      amount: '$50.00',
+      product: clothImg
+  },
+  {
+      id: 5,
+      brand: 'Apollo running shot',
+      amount: '$50.00',
+      product: clothImg
+  },
+  {
+      id: 6,
+      brand: 'Apollo running shot',
+      amount: '$50.00',
+      product: clothImg
+
+  },
+]
 
 const router = createBrowserRouter([
   {
@@ -15,16 +55,16 @@ const router = createBrowserRouter([
     children: [
       {
         path: '/',
-        element: <Women />,
-        
+        element: <Women cloth={cloth}/>,
+
       },
       {
         path: 'men',
-        element: <Men />,
+        element: <Men cloth={cloth}/>,
       },
       {
         path: 'kids',
-        element: <Kids />,
+        element: <Kids cloth={cloth}/>,
       },
     ]
   }
