+++ src/Pages/Men.js
@@ -1,11 +1,22 @@
 import React from 'react'
 import CategoryName from '../app/components/CategoryName'
+import ProductDescriptions from '../app/components/ProductDescriptions'
+import ProductImage from '../app/components/ProductImage'
 
-const Men = () => {
+const Men = ({cloth}) => {
+
+
+  const clothList = cloth.map(({id, brand, amount, product}) => (
+    <>
+      <ProductImage key={id} product={product}/>
+      <ProductDescriptions key={id} brand={brand} amount={amount} />
+    </>
+ ))
+
   return (
     <>
     <CategoryName />
-      <div>Men</div>
+      <div>{clothList}</div>
     </>
   )
 }
